[NumpyListVectorArrayMatrixOperator] move to pymor.operators.list
